requirements.txt
# Core Dependencies
faster-whisper>=0.10.0           # For Whisper speech recognition on CPU
edge-tts>=6.1.3                  # For Microsoft Edge text-to-speech
numpy>=1.21.0
# torch>=1.10.0 # Removed for sandbox environment
sounddevice>=0.4.6               # For recording audio
scipy>=1.7.0
soundfile>=0.12.1
webrtcvad>=2.0.10                # Voice Activity Detection
python-dotenv>=1.0.0             # Load environment variables

# Audio Processing
# pydub>=0.25.1 # Removed for sandbox environment
# SpeechRecognition>=3.8.1 # Removed for sandbox environment
pyaudio>=0.2.11                  # Required by SpeechRecognition

# LLM & LangChain
langchain>=0.1.16
langchain-community>=0.0.26
langchain-openai>=0.0.8

# Database
sqlalchemy>=2.0.0
pymysql>=1.0.2                   # For MySQL database connection (SQLAlchemy compatible)
mysql-connector-python>=8.0.0   # For direct MySQL querying in custom utilities

# Networking
requests==2.31.0                 # Used in DeepSeek API and other HTTP requests

# Web Framework
Flask[async]>=2.3.0              # For creating web frontend or async API
flask-cors>=4.0.0                # Enables CORS support for Flask

# Optional/Utility
tqdm                             # For showing progress bars (if used in future)


